authorization policy, resource operation authorization, created by id columns, nlog refractor

+++ HogwartsAPI/Services/WandService.cs
@@ -1,9 +1,12 @@
 ﻿using AutoMapper;
+using HogwartsAPI.Authorization;
 using HogwartsAPI.Dtos.WandDtos;
 using HogwartsAPI.Entities;
 using HogwartsAPI.Exceptions;
 using HogwartsAPI.Interfaces;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.EntityFrameworkCore;
+using HogwartsAPI.Enums;
 
 namespace HogwartsAPI.Services
 {
@@ -11,10 +14,14 @@ namespace HogwartsAPI.Services
     {
         private readonly HogwartDbContext _context;
         private readonly IMapper _mapper;
-        public WandService(HogwartDbContext context, IMapper mapper)
+        private readonly IUserContextService _userContext;
+        private readonly IAuthorizationService _authorizationService;
+        public WandService(HogwartDbContext context, IMapper mapper, IUserContextService userContext, IAuthorizationService authorizationService)
         {
             _context = context;
             _mapper = mapper;
+            _userContext = userContext;
+            _authorizationService = authorizationService;
         }
         public async Task<IEnumerable<WandDto>> GetAll()
         {
@@ -30,6 +37,8 @@ namespace HogwartsAPI.Services
         public async Task<int> Create(CreateWandDto dto)
         {
             var wand = _mapper.Map<Wand>(dto);
+            wand.CreatedById = _userContext.UserId;
+
             await _context.Wands.AddAsync(wand);
             await _context.SaveChangesAsync();
             return wand.Id;
@@ -42,12 +51,30 @@ namespace HogwartsAPI.Services
             {
                 throw new ForbidException("This wand has an owner");
             }
+
+            var authorizationResult = await _authorizationService.AuthorizeAsync(_userContext.User, wand,
+                new ResourceOperationRequirement(ResourceOperation.Delete));
+
+            if (!authorizationResult.Succeeded)
+            {
+                throw new ForbidException("You can't delete a wand you didn't add");
+            }
+
             await _context.Wands.Where(w => w.Id == id).ExecuteDeleteAsync();
         }
 
         public async Task Modify(int id, ModifyWandDto dto)
         {
             var wand = await GetWandById(id);
+
+            var authorizationResult = await _authorizationService.AuthorizeAsync(_userContext.User, wand,
+                new ResourceOperationRequirement(ResourceOperation.Update));
+
+            if (!authorizationResult.Succeeded)
+            {
+                throw new ForbidException("You can't modify a wand you didn't add");
+            }
+
             wand.Price = dto.Price;
             await _context.SaveChangesAsync();
         }
